scripts/cloud_functions/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-auth==2.23.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
